Fix period decoding. Vibrato + Volume slide should always use the values of previous vibrato.

+++ src/mods/mod/PatternReader.cpp
@@ -148,7 +148,7 @@ namespace mods
                {
                   mixedSample += channel->readAndConsumeNextSample(length);
                }
-             mixedSample /= channels.size();
+             mixedSample /= static_cast<double>(channels.size());
              return RLESample(mixedSample, length, false);
           }
